Add solutions to exercises in C++ Primer (5th edition)
After thinking for a while, I finally decide to add my own solutions to
the exercises in the fifth edition of C++ Primer into this repo.

I didn't add those codes in the past (although I have been working on
this book for quite a while) because I am not confident about my own
work. I am afraid that some people might googling into this repo and
take these guesswork as some sample solutions. As a beginner, my
knowledge of the programming language is very limited, and my solutions
to the exercises might be hopelessly wrong.

The codes I put here definately have nothing to do with "standard
solutions," no... These are just something I write for practices (and
for fun). They are very personal. Sometime I won't even follow the
direction, but do something in my own way. I put them here mostly as
notes for my study, but it is also great if you find the codes helpful.